requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
confuse==1.0.0
cycler==0.10.0
dremio-client==0.8.4
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
multidict==4.7.4
numpy==1.18.1
pandas==0.25.3
pyodbc==4.0.28
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
simplejson==3.17.0
six==1.14.0
slackclient==2.5.0
urllib3==1.25.7
Werkzeug==0.16.0
yarl==1.4.2
